+++ src/components/index.ts
@@ -10,3 +10,4 @@ export * from './mumble/MumblePost';
 export * from './mumble/RenderMumbles';
 export * from './navigation/NavigationComponent';
 export * from './alert/Alert';
+export * from './mumble/MumbleLike';
